% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy.R
\name{Accuracy}
\alias{Accuracy}
\title{Classification Accuracy
\code{Accuracy()} computes the percent identical values, controlling for perfect chance-level}
\usage{
Accuracy(obs, pred, na.rm = FALSE, response = c("discrete",
  "continuous"), ...)
}
\arguments{
\item{obs}{A numeric vector or matrix, the observed data. Can be continuous values or dicrete.  Can be aggregated, and if so you \strong{must} supply \code{n} (see below). The default assumes raw data.}

\item{pred}{A numeric vector or matrix with predictions, in the same order as \code{obs}.}

\item{na.rm}{(optional) Logical, \code{TRUE} means list-wise removal of \code{NA}, removes all \code{NA} rows found in \code{pred} or \code{obs} jointly.}

\item{response}{(default \code{"discrete"}) A string with response format, \code{"continuous"} or \code{"discrete"} (e.g., choices). Can be abbreviated. Will be inferred if missing from \code{obs}, \code{pred}.}

\item{...}{arguments from other methods, currently ignored.}
}
\value{
The percent of accurate predictions between 0 and 1, inclusive, where 1 indicates 100% accuracy.
}
\description{
Classification Accuracy
\code{Accuracy()} computes the percent identical values, controlling for perfect chance-level
}
\details{
Note: Accuracy corrects for perfect chance. Perfect chance predictions for discrete responses (e.g. 0.50 for binary responses) yield chance-level accuracy (50 %), rather than zero accuracy.
}
\examples{
Accuracy(obs = c(1,0,0), pred = c(1,1,1))        # \% accuracy
}
\seealso{
Other goodness of fit functions: \code{\link{APE}},
  \code{\link{MAPE}}, \code{\link{MDAPE}},
  \code{\link{MSE}}, \code{\link{RMSE}}, \code{\link{SSE}},
  \code{\link{gof}}
}
\concept{goodness of fit functions}
