% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APE.R
\name{APE}
\alias{APE}
\title{\code{APE()} calculates the average percentage error}
\usage{
APE(obs, pred, response = .guessResponse(obs, pred), ...)
}
\arguments{
\item{obs}{A numeric vector or matrix, the observed data. Can be continuous values or dicrete.  Can be aggregated, and if so you \strong{must} supply \code{n} (see below). The default assumes raw data.}

\item{pred}{A numeric vector or matrix with predictions, in the same order as \code{obs}.}

\item{response}{(default \code{"discrete"}) A string with response format, \code{"continuous"} or \code{"discrete"} (e.g., choices). Can be abbreviated. Will be inferred if missing from \code{obs}, \code{pred}.}

\item{...}{arguments from other methods, currently ignored.}
}
\value{
Absolute percentage error between \code{obs} and \code{pred}.
}
\description{
\code{APE()} calculates the average percentage error
}
\examples{
### APE
# --------------------------------------------------------------------------
# First example from Gilliland, M. (2010). 
# The business forecasting deal: exposing myths, eliminating bad practices,
# providing practical solutions (Vol. 27). John Wiley & Sons. (Exhibit A.3):
pred <- rep(1250, 6)
obs <- c(1150, 1400, 1200, 1350, 1100, 1300)
APE(obs, pred, response = "c")

obs <- diag(x = 1, ncol = 3, nrow = 3)
pred <- matrix(c(0.7, 0.2, 0.1,
                 0.3, 0.4, 0.3,
                 0.1, 0.1, 0.8), byrow = TRUE, nrow = 3)
APE(obs, pred, response = 'd')
obs <- 1:3
pred <- matrix(c(0.7, 0.2, 0.1,
                 0.3, 0.4, 0.3,
                 0.1, 0.1, 0.8), byrow = TRUE, nrow = 3)
APE(obs, pred, response = 'd')
                 
obs <- diag(x = 1, ncol = 3, nrow = 3)
pred <- c(0.7, 0.4, 0.8)
APE(obs, pred, response = 'd')
}
\seealso{
Other goodness of fit functions: \code{\link{Accuracy}},
  \code{\link{MAPE}}, \code{\link{MDAPE}},
  \code{\link{MSE}}, \code{\link{RMSE}}, \code{\link{SSE}},
  \code{\link{gof}}
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}, Florian Seitz
}
\concept{goodness of fit functions}
