% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicerule.R
\name{choicerule}
\alias{choicerule}
\title{Apply a choice rule (action selection rule) to probabilistic predictions.}
\usage{
choicerule(x, type = c("softmax", "luce", "argmax", "epsilon"),
  tolerance = sqrt(.Machine$double.eps), ...)
}
\arguments{
\item{x}{A numeric vector or matrix with probabilistic predictions for actions. If \code{x} is a vector, binary actions are assumed. If \code{x} is a matrix, column 1 holds predictions for action 1, column 2 for action 2, etc. \code{0 <= x <= 1} is required.}

\item{type}{The choice rule (action selection rule), can be one of \code{c("luce", "argmax", "softmax", "epsilon")}, abbreviations are allowed. \code{luce} is Luce's choice rule, \code{argmax} is an arg max choice rule, \code{softmax} is a soft maximum, \code{epsilon} is an epsilon-greedy rule assuming a constant error}

\item{tolerance}{**optional** A small number. Usually all rows in the predictions need to sum up to 1 if predictions contain more than 1 column. A \code{tolerance} allows the rows to sum up to 1 - tolerance, should be small.}

\item{...}{Other parameters to be passed to the choice rule methods: \code{softmax} needs a parameter \code{tau}, \code{epsilon} needs a parameter \code{eps}.}
}
\value{
A matrix holding the probability to select each action in the columns.
}
\description{
Apply a choice rule (action selection rule) to probabilistic predictions.
}
\examples{
# Some predictions
predictions <- cbind(A = c(.1,.5,.4), B = c(.3,.1,.4), C = c(.6, .4, .2))
# Apply various choice rules
\donttest{
choicerule(predictions, "luce")
choicerule(predictions, "argmax")
choicerule(predictions, "epsilon", eps = .2)

# Some binary predictions
binaryPredictions <- c(.22, .5, .73)
# Apply a choice rule
choicerule(binaryPredictions, "softmax", tau = 2)
}
}
\references{
See the individual functions.
}
