% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_argmax.R
\name{cr_argmax}
\alias{cr_argmax}
\title{Apply an arg max choice rule}
\usage{
cr_argmax(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A numeric vector or matrix with probabilistic predictions for actions. If \code{x} is a vector, binary actions are assumed. If \code{x} is a matrix, column 1 holds predictions for action 1, column 2 for action 2, etc.}
}
\value{
A numeric vector or matrix (depending on the input) of indicators of the maximum.
}
\description{
Apply an arg max choice rule
}
\details{
The returned values are typically 1 for the row maximum or 0, except if n actions have the maximum probability the result is 1/n, correcting ties by equal choice probability. If the iput is a vector it is silently coerced to a two-column matrix.
}
\examples{
# No examples
cr_argmax(c(0.1, 0.5, 0.8))                    # Returns a vector
cr_argmax(cbind(c(0.1, 0.9), c(0.5, 0.4)))      # Returns a matrix
}
\references{
Sutton, R. S., & Barto, A. G. (1998). Reinforcement learning: An introduction. Cambridge, MA: MIT Press
}
