% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_epsilon.R
\name{cr_epsilon}
\alias{cr_epsilon}
\title{Apply Epsilon greedy choice rule}
\usage{
cr_epsilon(x, eps)
}
\arguments{
\item{x}{A numeric vector or matrix with probabilistic predictions for actions. If \code{x} is a vector, binary actions are assumed. If \code{x} is a matrix, column 1 holds predictions for action 1, column 2 for action 2, etc.}

\item{eps}{A number between 0 an 1: the probability to make random choices across all options.}
}
\value{
A matrix holding the probability to select each action in the columns
}
\description{
Apply Epsilon greedy choice rule
}
\examples{
# No examples
}
\references{
Sutton, R. S., & Barto, A. G. (1998). Reinforcement learning: An introduction. Cambridge, MA: MIT Press
}
