% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_luce.R
\name{cr_luce}
\alias{cr_luce}
\title{Apply Luce choie rule, aka Clarke choice rule}
\usage{
cr_luce(x)
}
\arguments{
\item{x}{A numeric vector or matrix with probabilistic predictions for actions. If \code{x} is a vector, binary actions are assumed. If \code{x} is a matrix, column 1 holds predictions for action 1, column 2 for action 2, etc.}
}
\value{
A matrix holding the probability to select each action in a column
}
\description{
Apply Luce choie rule, aka Clarke choice rule
}
\examples{
# No examples
}
\references{
Luce, R. D. (1959). Individual Choice Behavior: A Theoretical Analysis. John Wiley and sons. New York.
Clarke, F. R. (1957). Constant‐ratio rule for confusion matrices in speech communication. The Journal of the Acoustical Society of America, 29(6), 715-720.
}
