% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_softmax.R
\name{cr_softmax}
\alias{cr_softmax}
\title{Apply soft maximum (softmax) choice rule for binary predictions}
\usage{
cr_softmax(x, tau)
}
\arguments{
\item{x}{A numeric vector or matrix with probabilistic predictions for actions. If \code{x} is a vector, binary actions are assumed. If \code{x} is a matrix, column 1 holds predictions for action 1, column 2 for action 2, etc. \code{x} can also be a value difference such as \code{x = price - utility}.}

\item{tau}{A number above 0 making action selection more random (aka temperature parameter). Large values make actions equiprobable, small values close to zero generate deterministic choices, close to arg max choices or Softmax-greedy choices.}
}
\value{
A matrix holding the probability to select each action in a column
}
\description{
Apply soft maximum (softmax) choice rule for binary predictions
}
\examples{
# No examples
}
\references{
Sutton, R. S., & Barto, A. G. (1998). Reinforcement learning: An introduction. Cambridge, MA: MIT Press
}
