% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loglikelihood.R
\name{Loglikelihood}
\alias{Loglikelihood}
\title{Log likelihood (log loss)}
\usage{
Loglikelihood(obs, pred, pdf = c("binomial", "normal", "truncnormal"),
  na.rm = FALSE, n = NULL, saturated = FALSE,
  eps = sqrt(.Machine$double.eps), binomial.coef = FALSE,
  sigma = NULL, a = NULL, b = NULL, ...)
}
\arguments{
\item{obs}{A numeric vector or matrix, the observed data. Can be continuous values or dicrete.  Can be aggregated, and if so you \strong{must} supply \code{n} (see below). The default assumes raw data.}

\item{pred}{A numeric vector or matrix with predictions, in the same order as \code{obs}.}

\item{pdf}{A string, probability density function;, allowed are \code{"binomial"}, \code{"normal"}, \code{"truncnormal"}. When using \code{"normal"} then \code{pred} needs two columns: observed means and observed standard deviations.}

\item{na.rm}{(optional) Logical. \code{TRUE} removes removes all \code{NA} rows found in \code{pred} or \code{obs} jointly (list-wise removal).}

\item{n}{(default: 1) integer or integer vector, number of observations underlying obs; required if \code{obs}, \code{pred} or both are \bold{aggregated}; \code{n=10} means each aggregate represents \code{10} data points, a vector \code{n=c(10,20)} means the first aggregate represents 10, the second 20 data points, etc.}

\item{saturated}{(optional) logical, compute saturated model (observed proportions as prediction, which approximates the highest possible likelihood). If so \code{obs} must be raw data, and \code{pred} can be omitted. Equal to \code{obs} being raw and \code{pred} being proportions.}

\item{eps}{(optional) small numeric to offset predictions from zero in  binomial pdf by substituting \code{log(eps)} for \code{log(0) = -Inf}. Don't change.}

\item{binomial.coef}{(optional) Logical (defaul is \code{FALSE}). If \code{TRUE} the binomial coefficient is used with \code{pdf="binomial"}. Mimicks results from \code{\link{dbinom}}. Ignored for other \code{pdf}s.}

\item{sigma}{(optional) Standard deviation of the normal distribution, \bold{required} for \code{pdf = "normal", "truncnormal"}).}

\item{a, b}{(optional) A number, lower/upper bound of the truncated pdf, required if \code{pdf = "truncnormal"}.}

\item{...}{Other parameters, currently ignored.}
}
\value{
The log likelihood of \code{obs} given  \code{pred}. Lokelihoods are computed without the binomial coefficient.
}
\description{
Log likelihood (log loss)
}
\examples{
\donttest{
# Example from Busemeyer & Diederich (2010)
# Observed relative frequencies of choice 1 from 11 conditions
obs <- c(.9538, .9107, .9204, .9029, .8515, .9197,
          .7970, .8228, .8191, .7277, .7276)
# Predictions for each of the 11 conditions
pred <- c(.9526, .9168, .8721, .8229, .7736, .7277,
          .6871, .6523, .6232, .5993, .5798)
n <- 200 # number of observations
# lnL in paper is -969.9514, < 0.1 \% deviation
ll <- Loglikelihood(obs = obs, pred = pred, n = 200)
# all.equal(ll, -969.9514, 0.001)
# lnLS in paper is -879.9013, < 1 \% deviation due to rounding
llsat <- Loglikelihood(obs = obs, n = 200, saturated = TRUE)
# all.equal(llsat, -879.9013, .01)

# Using the raw data
# Recreate the raw data (observations 0 or 1)
obsraw  <- rep(rep(0:1, 11), round(c(t(cbind(1-obs, obs))) * 200))
predraw <- rep(pred, each = 200)
ll <- Loglikelihood(obs = obsraw, pred = predraw)
# all.equal(ll, -969.9514, 0.001)
llsat <- Loglikelihood(obs = obsraw, pred = predraw, saturated = TRUE)
# all.equal(llsat, -879.9013, .01)
}
}
\references{
{ Busemeyer, J. R., & Diederich, A. (2010). Nonlinear parameter estimation. In Cognitive Modeling (pp. 43–84). Thousand Oaks, CAL: SAGE Publications. }
}
