% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDAPE.R
\name{MDAPE}
\alias{MDAPE}
\title{\code{MDAPE()} calculate median absolute percentage error}
\usage{
MDAPE(obs, pred, response = .guessResponse(obs, pred), ...)
}
\arguments{
\item{obs}{A numeric vector or matrix, the observed data. Can be continuous values or dicrete.  Can be aggregated, and if so you \strong{must} supply \code{n} (see below). The default assumes raw data.}

\item{pred}{A numeric vector or matrix with predictions, in the same order as \code{obs}.}

\item{response}{(default \code{"discrete"}) A string with response format, \code{"continuous"} or \code{"discrete"} (e.g., choices). Can be abbreviated. Will be inferred if missing from \code{obs}, \code{pred}.}

\item{...}{more arguments to be passed on to the fitting functions, see e.g. \code{\link[=loglikelihood]{loglikelihood()}}}
}
\value{
Median absolute percentage error between \code{obs} and \code{pred}
}
\description{
\code{MDAPE()} calculate median absolute percentage error
}
\examples{
# First example from Gilliland, M. (2010). 
# The business forecasting deal: exposing myths, eliminating bad practices,
# providing practical solutions (Vol. 27). John Wiley & Sons. (Exhibit A.3):
pred <- rep(1250, 6)
obs <- c(1150, 1400, 1200, 1350, 1100, 1300)
MDAPE(obs, pred, response = "con")
}
\seealso{
Other goodness of fit functions: 
\code{\link{APE}()},
\code{\link{Accuracy}()},
\code{\link{MAPE}()},
\code{\link{MSE}()},
\code{\link{RMSE}()},
\code{\link{SSE}()},
\code{\link{gof}()}
}
\author{
Jana B. Jarecki, \email{jj@janajarecki.com}, Florian Seitz
}
\concept{goodness of fit functions}
