% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansd.R
\name{meansd}
\alias{meansd}
\title{Compuate mean and (finite-sample corrected) standard deviation}
\usage{
meansd(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical, whether to remove \code{NA} values in \code{x}.}
}
\value{
A named list with the mean and the standard deviation
}
\description{
Compuate mean and (finite-sample corrected) standard deviation
}
\examples{
x <- sample(1:100)
msd(x)
}
