% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE.R
\name{MSE}
\alias{MSE}
\title{\code{MSE()} computes the mean squared error}
\usage{
MSE(obs, pred, na.rm = FALSE, weighted = FALSE, n = NULL, ...)
}
\arguments{
\item{obs}{A numeric vector or matrix, the observed data. Can be continuous values or dicrete.  Can be aggregated, and if so you \strong{must} supply \code{n} (see below). The default assumes raw data.}

\item{pred}{A numeric vector or matrix with predictions, in the same order as \code{obs}.}

\item{na.rm}{(optional) Logical. \code{TRUE} removes removes all \code{NA} rows found in \code{pred} or \code{obs} jointly (list-wise removal).}

\item{weighted}{(optional) Logical (dfault \code{FALSE}), if \code{TRUE} variance-weighted sums of squares are calculated, otherwise unweighted sums of squares.}

\item{n}{(optional), number of integer vector, required if \code{weighted = TRUE}. Number of observations underlying each entry in \code{obs}. One number if each entry in \code{obs} is the aggregate of the same number, a vector if different numbers of observations underly entries in code{obs}.}

\item{...}{other arguments are ignored}
}
\value{
Mean of squared deviations between \code{obs} and \code{pred}
}
\description{
\code{MSE()} computes the mean squared error
}
\examples{
# None so far
}
\seealso{
Other goodness of fit functions: \code{\link{APE}},
  \code{\link{Accuracy}}, \code{\link{MAPE}},
  \code{\link{MDAPE}}, \code{\link{RMSE}},
  \code{\link{SSE}}, \code{\link{gof}}
}
\concept{goodness of fit functions}
