% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSE.R
\name{SSE}
\alias{SSE}
\title{Sum of squared errors}
\usage{
SSE(obs, pred, na.rm = FALSE, weighted = FALSE, n = NULL, ...)
}
\arguments{
\item{obs}{A numeric vector or matrix, the observed data. Can be continuous values or dicrete.  Can be aggregated, and if so you \strong{must} supply \code{n} (see below). The default assumes raw data.}

\item{pred}{A numeric vector or matrix with predictions, in the same order as \code{obs}.}

\item{na.rm}{(optional) Logical. \code{TRUE} removes removes all \code{NA} rows found in \code{pred} or \code{obs} jointly (list-wise removal).}

\item{weighted}{(optional) Logical (dfault \code{FALSE}), if \code{TRUE} variance-weighted sums of squares are calculated, otherwise unweighted sums of squares.}

\item{n}{(optional), number of integer vector, required if \code{weighted = TRUE}. Number of observations underlying each entry in \code{obs}. One number if each entry in \code{obs} is the aggregate of the same number, a vector if different numbers of observations underly entries in code{obs}.}

\item{...}{Other parameters to be passed to the methods}
}
\value{
Sum of squared deviations between \code{obs} and \code{pred}, or weighted sum of squared deviations.
}
\description{
\code{SSE()} computes the sum of squared errors.
}
\examples{
# Example from Busemeyer & Diederich (2010), p. 55 - 56
# Observed relative frequencies of choices as proportions of choice 1, split up by 11 conditions
obs <- c(.9538, .9107, .9204, .9029, .8515, .9197,
          .7970, .8228, .8191, .7277, .7276)
# Predictions for each of the 11 conditions
pred <- c(.9526, .9168, .8721, .8229, .7736, .7277,
          .6871, .6523, .6232, .5993, .5798)
n <- 200 # all  or n = rep(200, 11) # number of observations per condition

sse <- SSE(obs = obs, pred = pred)
# Sum of squared errors (SSE) in paper equals 0.1695
# all.equal(sse, 0.1695, 0.001) # TRUE

wsse <- SSE(obs = obs, pred = cbind(pred, 1-pred), weighted = TRUE, n = 200)
# Weighted sum of squared errors (WSSE) in paper equals 158.4059
# all.equal(wsse, 158.4059, 0.001) TRUE
# minute difference is due to rounding
}
\references{
Busemeyer, J. R., & Diederich, A. (2010). Nonlinear parameter estimation. In Cognitive Modeling (pp. 43–84). Thousand Oaks, CAL: SAGE Publications.
}
\seealso{
Other goodness of fit functions: 
\code{\link{APE}()},
\code{\link{Accuracy}()},
\code{\link{MAPE}()},
\code{\link{MDAPE}()},
\code{\link{MSE}()},
\code{\link{RMSE}()},
\code{\link{gof}()}
}
\concept{goodness of fit functions}
