% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansd.R
\name{print.meansd}
\alias{print.meansd}
\title{Computes the mean and standard deviation and formats them as string}
\usage{
\method{print}{meansd}(x, digits = 2, na.rm = TRUE, label = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{The number of digits to print}

\item{label}{(default \code{TRUE}) Logical, if true M and SD are added as labels}
}
\value{

}
\description{
Computes the mean and standard deviation and formats them as string
}
